+++ shardingsphere-sql-rewriter/shardingsphere-sql-rewriter-engine/src/main/java/org/apache/shardingsphere/sql/rewriter/sql/token/pojo/generic/InsertValuesToken.java
@@ -18,16 +18,9 @@
 package org.apache.shardingsphere.sql.rewriter.sql.token.pojo.generic;
 
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.sql.rewriter.sql.token.pojo.SQLToken;
 import org.apache.shardingsphere.sql.rewriter.sql.token.pojo.Substitutable;
-import org.apache.shardingsphere.core.rule.DataNode;
-import org.apache.shardingsphere.sql.parser.sql.segment.dml.expr.ExpressionSegment;
-import org.apache.shardingsphere.sql.parser.sql.segment.dml.expr.complex.ComplexExpressionSegment;
-import org.apache.shardingsphere.sql.parser.sql.segment.dml.expr.simple.LiteralExpressionSegment;
-import org.apache.shardingsphere.sql.parser.sql.segment.dml.expr.simple.ParameterMarkerExpressionSegment;
 
-import java.util.Collection;
 import java.util.LinkedList;
 import java.util.List;
 
@@ -42,42 +35,11 @@ public abstract class InsertValuesToken extends SQLToken implements Substitutabl
     
     private final int stopIndex;
     
-    private final List<InsertValueToken> insertValueTokens;
+    private final List<InsertValue> insertValues;
     
     public InsertValuesToken(final int startIndex, final int stopIndex) {
         super(startIndex);
         this.stopIndex = stopIndex;
-        insertValueTokens = new LinkedList<>();
-    }
-    
-    @RequiredArgsConstructor
-    @Getter
-    public static final class InsertValueToken {
-        
-        private final List<ExpressionSegment> values;
-        
-        private final Collection<DataNode> dataNodes;
-        
-        @Override
-        public String toString() {
-            StringBuilder result = new StringBuilder();
-            result.append("(");
-            for (int i = 0; i < values.size(); i++) {
-                result.append(getValue(i)).append(", ");
-            }
-            result.delete(result.length() - 2, result.length()).append(")");
-            return result.toString();
-        }
-        
-        private String getValue(final int index) {
-            ExpressionSegment expressionSegment = values.get(index);
-            if (expressionSegment instanceof ParameterMarkerExpressionSegment) {
-                return "?";
-            } else if (expressionSegment instanceof LiteralExpressionSegment) {
-                Object literals = ((LiteralExpressionSegment) expressionSegment).getLiterals();
-                return literals instanceof String ? String.format("'%s'", ((LiteralExpressionSegment) expressionSegment).getLiterals()) : literals.toString();
-            }
-            return ((ComplexExpressionSegment) expressionSegment).getText();
-        }
+        insertValues = new LinkedList<>();
     }
 }
